+++ packages/agtree/src/index.ts
@@ -7,6 +7,7 @@ export { RuleParser } from './parser/rule-parser';
 export { RuleSerializer } from './serializer/rule-serializer';
 export { RuleDeserializer } from './deserializer/rule-deserializer';
 export { RuleGenerator } from './generator/rule-generator';
+export { type OnParseError, defaultParserOptions, type ParserOptions } from './parser/options';
 
 export {
     type Agent,